web/requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.2
django-appconf==1.0.2
django-compressor==2.2
django-registration==3.0
django-sass-processor==0.7.1
factory-boy==2.11.1
Faker==0.9.1
gunicorn==19.9.0
libsass==0.15.1
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
text-unidecode==1.2
